Move files from redux folder to components folder
