+++ src/components/BookItem.js
